Fix mypy --strict errors in M8 velocity/mass recovery tests
The M8 test suite tripped mypy under --strict (17 errors across 5 files);
CI lint-and-test failed on the mypy leg. Fixes are test-only:

- test_velocity_recovery.py: annotate the _convert_done_criterion helper
  with its real ConversionResult return type (was object, so every
  .report/.validation/.output access failed attr-defined), and narrow
  ConversionResult.output (bytes | None) before passing it to _parse.
- test_missing_velocities/masses.py, test_poscar.py: narrow the optional
  canonical fields (velocities/masses, ndarray | None) with `is not None`
  before assert_allclose / indexing.
- test_physics_maxwell_boltzmann.py: wrap the ASE-units expression (typed
  Any, since ase is follow_imports=skip) in float() to satisfy the
  declared return type.

No production code touched; behavior unchanged (49 affected tests pass).

+++ tests/recovery/test_physics_maxwell_boltzmann.py
@@ -16,7 +16,7 @@ from xtalate.recovery.engine import _maxwell_boltzmann
 
 def _expected_variance(mass_u: float, temperature_k: float) -> float:
     # σ² = kB·T/m in ASE velocity units, scaled to (Å/fs)² by ase.units.fs² (the codebase's factor).
-    return ase_units.kB * temperature_k / mass_u * ase_units.fs**2
+    return float(ase_units.kB * temperature_k / mass_u * ase_units.fs**2)
 
 
 def test_per_component_variance_matches_kT_over_m() -> None:
